Batch writer to main for real this time (#2017)
## Description

I did a a PR to merge the batch writer change. But it somehow got lost.
https://github.com/n0-computer/iroh/pull/2007

It was targeted to
[fallible-store-traits](https://github.com/n0-computer/iroh/pull/2005),
which got merged to main. I would have expected github to retarget the
PR to main, but apparently it did not do that. Maybe I made a mistake,
not sure exactly.

So here we go again.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.