fix(iroh-net): Fix in detecting globally routable IPv6 addresses (#2030)
## Description

We were incorrectly identifying globally routable IPv6 addresses as
not routable.  This make netcheck skip IPv6 addresses.

## Notes & open questions

Closes #2022 

The fix in the netcheck, test is curious, it fails on this PR fairly
consistently yet it is really a flaky test introduced in combination
with #2027.
Now that we have IPv6 resolving properly we might get an IPv6 result.
Since
this test only has a single DerpUrl it only needs one result to
complete.
This could be an IPv4 OR an IPv6 result, we don't know which one will be
faster. And if the other is slow enough then it won't be received. I
think
this behaviour is fine, we simply can't guarantee that both must have a
result
and the point of netcheck is to return something working quickly.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.