fix(iroh-net): Handle unreachable IPv6 networks better (#2029)
## Description

Now that we start supporting IPv6 we see this a lot more.  If we have
a dual-stack machine it does not mean the IPv6 network is routable.

If the network is not routable there is no point in trying more
attempts in the probeset, this just delays the failure.  This results
in a visible speed up on my local machine.

## Notes & open questions

Tested on routed and unrouted dual-stack machines

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.