Introduce BaoBlobSize enum
for now this is an enum with a size and verified/unverified. In the future
we might add more info to the unverified path, such as the max offset
we have validated. But this info is not always available easily, so we
don't for now.

If you just want the size value, there is .value().