refactor(iroh-bytes): Simplify store traits (#2023)
## Description

Simplify and rename some store traits, taking advantage of return
position impl trait in trait.

The main change is that the map entries are no longer for one specific
map impl, but just generic map entries, so the type parameter goes away.
Other than that, the PartialXXX traits are renamed to XXXMut, which is
more accurate. An immutable entry can be partial. Since MapEntry does
not guarantee that it is complete (anymore), the main difference between
MapEntry and MapEntryMut is not that the latter allows writing.

Edit: one additional change is to introduce BaoBlobSize, which is a size
together with an information about whether the size is validated or not.
Why not just (u64, bool)? Because later we might add more info to the
unverified enum case.

## Notes & open questions

Note: this is not the end of it. Just trying to do this incrementally.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] Tests if relevant.