refactor(iroh-net): Improve API to retrieve local endpoints
The API to retrieve local endpoints was split in two between the
current endpoints and updates to it.  This was very hard to use right
without any race conditions.

This replaces it with a single API which returns a stream.  This is a
little bit wordier to use if you just need the current local item, but
much harder to misuse.