refactor(iroh-net): Improve API to retrieve local endpoints (#2041)
## Description

The API to retrieve local endpoints was split in two between the
current endpoints and updates to it.  This was very hard to use right
without any race conditions.

This replaces it with a single API which returns a stream.  This is a
little bit wordier to use if you just need the current local item, but
much harder to misuse.

## Notes & open questions

It would be possible to add a convenient `.current_local_endpoints()`
API if we wanted.  Not sure this is really worth it.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.