fix(iroh-net): Trigger netcheck on a magicsock rebind (#2042)
## Description

This incorporates two fixes:

- The .set_preferred_port() function was completely ignoring the passed
port. This is now plugged through to the rebind. Nobody was using this
function before though. I'm not so sure how useful it is.

- When rebinding the locally bound sockets change. This means we have to
do a new netcheck as otherwise we won't know about our new endpoints
until a new netcheck runs on it's regular schedule.

## Notes & open questions

This is kind of split off from #2041 because it ends up doing unrelated
fixes just for an improvement to tests.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.