chore(ci): Warn in our discord channel on failure (#2044)
## Description

This is better than error, since flaky tests really are expected to be
flaky.  And we don't want to miss other errors in that channel which
are more important.

## Notes & open questions

Maybe warn is even a bit high, but all that matters is that we get
different colours in our channel.  If later we send more notifications
to that channel we can tweak the levels appropriately.

## Change checklist

- [x] Self-review.