refactor(iroh-net): remove unneeded async interactions with the magicsock actor (#2058)
## Description

While working on #2056 I spotted that we use the actor inbox with return
channels for information that is readily available already on the shared
inner magicsock. This removes the unneeded complexity and thus
simplifies `get_mapping_addr`, `endpoint_info` and `endpoint_infos` to
return the information non-async and infallible. Yay!

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.