feat(iroh-net): Use the local endpoints info when closing derps
When the network has changed we used to close all the derp connections
unconditionally.  This passes in the local endpoints information so
that the derp http actor will only close those derp connections for
which we no longer have the local endpoint.  That's a lot less
disruptive than always closing them.