feat(iroh-net): Use the local endpoints info when closing derps (#2082)
## Description

When the network has changed we used to close all the derp connections
unconditionally.  This passes in the local endpoints information so
that the derp http actor will only close those derp connections for
which we no longer have the local endpoint.  That's a lot less
disruptive than always closing them.

## Notes & open questions

The relevant rebinding test is still marked flaky.  I've only seen it
pass but will monitor in the flaky tests before considering enabling.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.