refactor(iroh-net): remove rebinding (#2083)
## Description

This removes the rebinding functionality of the magicsocket. We did not
use this functionality, it was incoherently implemented and thus very
confusing what of it was still relevant or not. While there is some
theoretical need for rebinding, it is cleaner to remove it all. The main
casualty of this is the ability to set a preferred port, to this the
answer is to ask to bind to the preferred port to start with, in the
EndpointBuilder. The other casualties are some test which now can not be
used anymore, they're not a great loss.

We still keep the behaviour of reacting to major network changes by
resetting node path state and by pruning derp connections.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.