refactor(redb-store): Optimization for small file import in redb store (#2062)
## Description

Optimization for small file import in redb store.

This batches writes and reads up to some delay and vastly improves
performance especially for writes. Linux kernel import goes from minutes
to seconds.

## Notes & open questions

~~Currently delete is just a normal operation. I think it should be
changed to a toplevel operation because crashing with a delete
transaction being not committed would lead to inconsistent state: files
would be gone on disk and there in the db.~~

File deletions of any of the "owned" files are now handled outside the
transaction. That requires keeping track of deletes and undeletes within
a transaction. E.g. if within a single transaction it is determined that
hash 0xabcd is to be deleted, it gets marked for deletion, then it gets
created again. We must **not** delete the file at the end of the
transaction.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] Tests if relevant.