refactor(iroh-bytes): Remove the size argument from get_or_create
We have *a* size, but we don't know if it is the correct size. At the point
where we call get_or_create it could be anything. So we must not use this
size for *anything* whatsoever, so we might as well omit it.

The purpose of the size is solely to make sense of the following data,
nothing else.