refactor(iroh-bytes): get rid of meta dir (#2111)
## Description

refactor(iroh-bytes): get rid of meta dir

blobs.db now lives on top level. The meta dir used to exist because we
had to store metadata in the file system. Now we don't have to do this
anymore, so the meta dir containing a single file seems a bit
superfluous.

cc @ppodolsky if we merge this you would have to manually move the blobs
file one level up for the places where you have already done an import.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.