refactor(iroh-net): allow to set a custom DNS resolver on the magic endpoint (#2116)
## Description

This makes the DNS resolver to be used in the context of a MagicEndpoint
configurable. Up to now, we used a single, global, unconfigurable DNS
resolver, stored in a per-process global static. This PR changes this so
that we can set a DNS resolver in the builder of the MagicEndpoint. This
resolver is passed through to all places where we need a DNS resolver -
which is all places where we need to resolve relay URLs.

The default is unchanged: A single, shared DNS resolver is used for all
endpoints. However, this default can now be changed per-endpoint. The
global resolver is only used as a default in the endpoint builder if no
custom resolver is set, and in the doctor, and in tests.

This change will make testing things that use DNS - prominently: #2045 -
much easier. And we now have the means in place for people to customize
the DNS resolving, if needed.

## Notes & open questions

This makes the `hickory_resolver::TokioAsyncResolver` part of the public
API surface of `iroh-net`.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] Tests if relevant.