fix: do not allow connecting to ourself (#2123)
## Description

This changes the `MagicEndpoint::connect` method to always return an
error if attempting to connect to ourselves (i.e. connecting to the
node_id of the node initiating the connection).

There is no use case for this, and it is better to surface the error
quickly than having users deal with weird situations in their apps later
on.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.