refactor(iroh-sync): rip out the mem implementation of the doc store (#2112)
## Description

refactor(iroh-sync): rip out the mem implementation of the doc store

This would be a prerequisite for doing major changes on how docs work.
Having 2 implementations that need to be kept in sync would slow that
down way too much.

## Notes & open questions

How costly is creating a mem redb instance? Is this a big deal? Are we
creating a mem instance and then throwing it away in the builder, or
not?

Todo: remove the tests associated with the mem store.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.