refactor(metrics): metrics on by default (#2129)
## Description

Now metrics default to serving on `localhost:9090` but you can override
with a custom port or turn off if you like with `--metrics-port`. If you
want to actually expose the metrics endpoint to the world, it's assumed
you will set up a config file with a custom `metrics_addr`.

I guess the config override could have been done in a bit more elegant
way but didn't want to commit refactoring the `NodeConfig` nor did I
want fancy extra task specific fn's.

The configuration is limited to the iroh node start path, the rest are
ignored as they are expected to be short lived processes.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.