fix(iroh): do not shut down node on internal rpc error (#2158)
## Description

fix(iroh): do not shut down node on internal rpc error

Internal rpc errors can happen, e.g. when the client side gets dropped.
No reason to shut down the entire node. We already have a cancellation
token, so it's not like we rely on this for shutdown.

Fixes #2157, #2143

## Notes & open questions

Note: I also sneaked in a fix for #2143. Basically when we are unable to
notify the db that a new gc epoch has started, we just shot down the gc
loop, since clearly the store is not well (probably shut down).

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.