feat(iroh-net): add `MagicEndpoint::wait_until_direct_conn`
This blocking method will return once a direct connection has been established between the endpoint and the node at `node_id`, or return in error if a timeout occurs. If no `Endpoint` exists for this `node_id` in the `NodeMap`, one will be created. This can be called before or after `connect`. If a direct connection already exists between the endpoint and the node at `node_id` when this method is called, it will immediately return.