fix(iroh-cli): Fix printing of doctor connect/accept output
Apparently now metrics work and are attempted to be printed.
Unfortunately they also print a tracing error message to stderr, the
stderr which is also being used by the progress bar.  The result is
that the progress bar ends up with an extra newline and the first line
of it is not being repainted but flows up, resulting in lots of
garbage in the terminal history.

By removing the log message the output is perserved much better.

It could be argued that this is not a good fix, if the loglevel is
changed to info or debug the same problems would occur.  However we
already have the ability to log to a random filedescriptor and will
soon be able to log to files.  So maybe that's just fine for now.