refactor(iroh-bytes): use even newer bao-tree (#2168)
## Description

refactor(iroh-bytes): use even newer bao-tree

There is a new version of bao-tree that fixes some long standing api
consistency issues and drops the tokio dependency. This updates iroh to
use it.

We of course still use tokio, but switching to something else would be
easier at least as far as bao-tree is concerned.

The other change is that encode_ranges_validated no longer writes the
length prefix, so we have to do this ourselves. This was done to
simplify the bao-tree API and make it more flexible.

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.