feat(iroh-net): add `MagicEndpoint::conn_type_stream` returns a stream that reports connection type changes for a `node_id` (#2161)
## Description
`MagicEndpoint::conn_type_stream` returns a `Stream` that reports
changes for a `magicsock::Endpoint` with a given `node_id` in a
`magicsock::NodeMap`.

It will error if no address information for that `node_id` exists in the
`NodeMap`.

This PR also adjusts the `Endpoint::info()` method to use the same
`ConnectionType` that gets reported to the stream.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.