feat: release automation (#2214)
## Description

This is a bit hard to test without merging. However merging should be
safe as it does nothing unless we trigger a release via push. We can
also toy with it then to see if things work.

Moving forward, pushing with a tag matching the pattern `v*` will
automatically create a release titled with the tag name and attach all
binaries (`iroh`, `iroh-relay`, `iroh-dns-server`) for all platforms we
support. Once done you can edit the release to update the title and
contents.

Also coveres doing `musl` builds for linux platforms.

Closes: https://github.com/n0-computer/iroh/issues/2174 &
https://github.com/n0-computer/iroh/issues/2173

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.