refactor(iroh-net): Rename endpoint for nodes to node_state
We still have many things named "endpoint", this targets cleaning up
the naming of the state for each node that is stored in the NodeMap.
This is now called NodeState instead of Endpoint.  All related APIs
now talk about nodes and node states instead of endpoints.

Another minor cleanup is in the NodeState we had private fields with
accessor functions, except for one field which was directly
accessible.  This is migrated to accessor functions for consistency as
well.