fix(iroh_net): less agressive best_addr clearing on pong timeout (#2238)
## Description

Clears the best address only if within the time the pong should have
been received, nothing else was seen so that the address is considered
not alive in that period. This should result in less aggressive
best_addr clearing.

Specific changes include:
- remove `BestAddr::clear_if_addr_older` since it is no longer needed.
- change `BestAddr::clear*` return since it was never read anyway.
- check for alive withing the `PING_TIMEOUT_DURATION` before clearing
the best address on timeout
- misc docs fixes in code related to being alive/active

## Breaking Changes

n/a

## Notes & open questions

n/a

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [ ] ~Tests if relevant.~
- [ ] ~All breaking changes documented.~