feat(doctor): Report connection type changes in rolling fashion
The doctor already shows the current connection type, this changes to
show changes in a rolling fashion.

This is not exact and some switches might be missed since the doctor
mechanism this hooks into is poll based.  A better version would be
using the new conn_type_stream and print each change.