feat(iroh-base): allow the addr info of tickets to be empty (#2254)
## Description

feat(iroh-base): allow the addr info of tickets to be empty

Without this, a ticket that does not contain any addresses or relay URL
can not be parsed. But we need to allow this if we want to use short
tickets.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

Note: I did not remove the error type Validation. That way we don't have
breaking changes, and we might need it again in the future.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.