refactor(iroh-net)!: Make the interfaces module private (#2266)
## Description

This is leaking 3rd party types into our public interface, since this
really only exists for our own use it probably shouldn't be public to
start with.  So make it private.  As a result cleans up some dead
code.

## Breaking Changes

The `iroh_net::net::interfaces` module has been made private.


## Notes & open questions

There is probably more in `iroh_net::net` that needs to be private.
Maybe even the whole module.  I'm not sure if doing this stepwise or
in one large chunk is better.  This takes the stepwise approach.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- [x] Tests if relevant.
- [x] All breaking changes documented.