feat(cli): add metrics server to iroh doctor (#2292)
## Description

Folks try to look at metrics while doctor alone is running (note this
clashes by default with the iroh node if both are running locally on
port :9090). Also kind of inconvenient given if you want to look at a
running iroh node when doctor is running.

Suggestions welcome, maybe default off metrics on `doctor`?

Also most metrics are not really used in the `doctor` path, but we can
fix that as we figure out what's cool to measure.

## Breaking Changes

<!-- Optional, if there are any breaking changes document them,
including how to migrate older code. -->

## Notes & open questions

<!-- Any notes, remarks or open questions you have to make about the PR.
-->

## Change checklist

- [x] Self-review.
- [ ] Documentation updates if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.