docs(iroh-net): Update toplevel module documentation (#2329)
## Description

This updates the top-level docs for iroh-net to be a general
overview of what the crate provides, since this is effectively the
docs.rs landing page.

## Breaking Changes

None

## Notes & open questions

This should include an examples section as well.  However... the API
is a bit pants and does not lend itself to a nice echo example here.
Let's tackle this separately.

This does not change the README, that is already fairly decent and
seems compatible with this content.

## Change checklist

- [x] Self-review.
- [x] Documentation updates if relevant.
- ~~[ ] Tests if relevant.~~
- ~~[ ] All breaking changes documented.~~

---------

Co-authored-by: Divma <26765164+divagant-martian@users.noreply.github.com>